Replace "Brocfile.js" with "ember-cli-build.js"
