+++ ember-cli-build.js
@@ -1,0 +1,158 @@
+var path = require('path');
+var resolve = require('resolve');
+var concat     = require('broccoli-concat');
+var Funnel     = require('broccoli-funnel');
+var mergeTrees = require('broccoli-merge-trees');
+var jshintTree = require('broccoli-jshint');
+var replace    = require('broccoli-string-replace');
+var gitVersion = require('git-repo-version');
+var BabelTranspiler = require('broccoli-babel-transpiler');
+
+function compileES6(tree) {
+  return new BabelTranspiler(tree, {
+    loose: true,
+    moduleIds: true,
+    modules: 'amdStrict'
+  });
+}
+
+module.exports = function(defaults) {
+  // --- Compile ES6 modules ---
+
+  var loader = new Funnel('bower_components', {
+    srcDir: 'loader',
+    files: ['loader.js'],
+    destDir: '/assets/'
+  });
+
+  var emberTestHelpersPath = path.dirname(resolve.sync('ember-test-helpers'));
+  var klassyPath = path.dirname(resolve.sync('klassy', { basedir: emberTestHelpersPath }));
+
+  var emberTestHelpers = new Funnel(emberTestHelpersPath, {
+    srcDir: '/',
+    include: [/\.js$/],
+    destDir: '/'
+  });
+
+  // TODO - this manual dependency management has got to go!
+  var klassy = new Funnel(klassyPath, {
+    srcDir: '/',
+    files: ['klassy.js'],
+    destDir: '/'
+  });
+  var deps = mergeTrees([klassy, emberTestHelpers]);
+
+  var lib = new Funnel('lib', {
+    srcDir: '/',
+    include: [/.js$/],
+    destDir: '/'
+  });
+
+  var tests = new Funnel('tests', {
+    srcDir: '/',
+    include: [/test-support\/resolver.js$/, /.js$/],
+    destDir: '/tests'
+  });
+
+  var main = mergeTrees([deps, lib]);
+  main = concat(compileES6(main), {
+    inputFiles: ['**/*.js'],
+    outputFile: '/ember-mocha.amd.js'
+  });
+
+  var generatedBowerConfig = new Funnel('build-support', {
+    srcDir: '/',
+    destDir: '/',
+    files: ['bower.json']
+  });
+  generatedBowerConfig = replace(generatedBowerConfig, {
+    files: ['bower.json'],
+    pattern: {
+      match: /VERSION_PLACEHOLDER/,
+      replacement: function() {
+        // remove leading `v` (since by default our tags use a `v` prefix)
+        return gitVersion().replace(/^v/, '');
+      }
+    }
+  });
+
+  var buildExtras = new Funnel('build-support', {
+    srcDir: '/',
+    destDir: '/',
+    files: ['README.md', 'LICENSE', 'mocha-setup.js']
+  });
+
+  var globalizedBuildSupport = new Funnel('build-support', {
+    srcDir: '/',
+    files: ['iife-start.js', 'globalize.js', 'iife-stop.js'],
+    destDir: '/'
+  });
+
+  var globalizedMain = concat(mergeTrees([loader, main, globalizedBuildSupport]), {
+    inputFiles: ['iife-start.js', 'assets/loader.js', 'ember-mocha.amd.js', 'globalize.js', 'iife-stop.js'],
+    outputFile: '/ember-mocha.js'
+  });
+
+  var jshintLib = jshintTree(lib);
+  var jshintTest = jshintTree(tests);
+
+  var mainWithTests = mergeTrees([deps, lib, tests, jshintLib, jshintTest]);
+  mainWithTests = concat(compileES6(mainWithTests), {
+    inputFiles: ['**/*.js'],
+    outputFile: '/assets/ember-mocha-tests.amd.js'
+  });
+
+  // --- Select and concat vendor / support files ---
+
+  var vendor = concat('bower_components', {
+    inputFiles: [
+      'jquery/dist/jquery.js',
+      'ember/ember-template-compiler.js',
+      'ember/ember.js',
+      'pretender/pretender.js',
+      'ember-data/ember-data.js'],
+    outputFile: '/assets/vendor.js'
+  });
+
+  var mocha = new Funnel('bower_components', {
+    srcDir: '/mocha',
+    files: ['mocha.js', 'mocha.css'],
+    destDir: '/assets'
+  });
+
+  var chai = new Funnel('bower_components', {
+    srcDir: '/chai',
+    files: ['chai.js'],
+    destDir: '/assets'
+  });
+
+  var adapter = new Funnel('bower_components', {
+    srcDir: '/ember-mocha-adapter',
+    files: ['adapter.js'],
+    destDir: '/assets'
+  });
+
+  var appShims = new Funnel('bower_components', {
+    files: ['ember-cli-shims/app-shims.js'],
+  });
+
+  var testLoader = new Funnel('tests', {
+    files: ['test-support/test-loader.js'],
+  });
+
+  var testSupport = concat(mergeTrees([appShims, testLoader]), {
+    inputFiles: [
+      'ember-cli-shims/app-shims.js',
+      'test-support/test-loader.js',
+    ],
+    outputFile: '/assets/test-support.js'
+  });
+
+  var testIndex = new Funnel('tests', {
+    srcDir: '/',
+    files: ['index.html'],
+    destDir: '/tests'
+  });
+
+  return mergeTrees([loader, main, mainWithTests, globalizedMain, vendor, mocha, chai, adapter, testSupport, testIndex, generatedBowerConfig, buildExtras]);
+};
